// src/types/product.ts
export interface ProductConfig {
  __no: number;
  __name: string;
  __fnsku: string;
  __asin: string;
  __sku: string;
  __ads: string[];
  __coupon: string[];
  __default_extra_purchase_price: string;
  __default_extra_weight: string;
  __default_extra_inside_express_price: string;
  __default_extra_shipping_price: string;
  __category: string;
}